added validation and crud for entry

+++ src/main/java/ch/zli/m223/punchclock/controller/EntryController.java
@@ -6,9 +6,11 @@ import ch.zli.m223.punchclock.domain.User;
 import ch.zli.m223.punchclock.dto.EntryDto;
 import ch.zli.m223.punchclock.service.EntryService;
 import ch.zli.m223.punchclock.service.UserService;
+import ch.zli.m223.punchclock.validator.EntryValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -23,6 +25,10 @@ public class EntryController {
     private final EntryService entryService;
     private final EntryDtoConverter entryDtoConverter;
     private final UserService userService;
+    private final EntryValidator entryValidator;
+
+    @InitBinder("entryDto")
+    public void setEntryDtoBinder(WebDataBinder binder){binder.setValidator(entryValidator);}
 
 
     @GetMapping("/api/entries")
@@ -40,10 +46,24 @@ public class EntryController {
         return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
-    @RequestMapping(value = "/entry/{id}", method = RequestMethod.DELETE)
-    public ResponseEntity<?> deleteEntry(@PathVariable("id") Long id){
+    @RequestMapping(value = "/api/entry", method = RequestMethod.PUT)
+    public ResponseEntity<?> updateEntry(@Valid @RequestBody EntryDto entryDto, Principal principal){
+        User user = userService.getByUsernameOrElseThrow(principal.getName());
+        if(entryDto.getId() != null && entryDto.getId() != 0){
+            Entry entry = entryDtoConverter.toEntity(entryDto, user);
+            Optional<Entry> oldEntry = entryService.findById(entryDto.getId());
+            if(oldEntry.isPresent() && oldEntry.get().getUser().equals(user)){
+                return new ResponseEntity<>(entryDtoConverter.toDto(entryService.updateEntry(oldEntry.get(), entry)), HttpStatus.OK);
+            }
+        }
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+    }
+
+    @RequestMapping(value = "/api/entry/{id}", method = RequestMethod.DELETE)
+    public ResponseEntity<?> deleteEntry(@PathVariable("id") Long id, Principal principal){
+        User user = userService.getByUsernameOrElseThrow(principal.getName());
         Optional<Entry> oldEntry = entryService.findById(id);
-        if(oldEntry.isPresent()){
+        if(oldEntry.isPresent() && oldEntry.get().getUser().equals(user)){
             entryService.deleteEntry(oldEntry.get());
             return new ResponseEntity<>(HttpStatus.OK);
         }else {
